Use open syscall for loop devices
